+++ omf-example-plugin/src/main/java/com/samares_engineering/omf/omf_example_plugin/features/cloneexample/actions/CloneType.java
@@ -15,7 +15,7 @@ import com.samares_engineering.omf.omf_core_framework.feature.registrables.actio
 import com.samares_engineering.omf.omf_core_framework.feature.registrables.actions.annotations.DeactivateListener;
 import com.samares_engineering.omf.omf_core_framework.feature.registrables.actions.annotations.DiagramAction;
 import com.samares_engineering.omf.omf_core_framework.feature.registrables.actions.annotations.MDAction;
-import com.samares_engineering.omf.omf_core_framework.utils.Clone.CloneManager;
+import com.samares_engineering.omf.omf_core_framework.utils.clone.CloneManager;
 import com.samares_engineering.omf.omf_core_framework.utils.OMFUtils;
 
 import java.util.List;